requirements.txt
flask
chromadb
sentence-transformers
pymysql
python-dotenv
genai
requests
xarray
numpy
pandas
sqlalchemy
glob